requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
docopt==0.6.2
fonttools==4.38.0
graphviz==0.20.1
gym==0.25.1
gym-notices==0.0.8
gym-pull==0.1.7
gym-super-mario-bros==7.4.0
idna==3.4
joblib==1.3.2
kiwisolver==1.4.4
matplotlib==3.6.3
neat-python==0.92
nes-py==8.2.1
numpy==1.24.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pipreqs==0.4.11
pyglet==1.5.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
scipy==1.11.2
six==1.16.0
termcolor==2.3.0
tqdm==4.66.1
urllib3==1.26.14
yarg==0.1.9
